backend/requirements.txt
pandas>=1.0.0
matplotlib>=3.0.0
python-dotenv>=0.19.0
python-dateutil>=2.8.0
pytest>=7.0.0
pytest-cov>=4.0.0
# Agentic Task & Management
firebase-admin>=6.0.0
google-generativeai>=0.3.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
